+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/RestrictedMemberChatMember.kt
@@ -5,7 +5,6 @@ package dev.inmo.tgbotapi.types.chat.member
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.chat.ChatPermissions
 import dev.inmo.tgbotapi.types.chat.PreviewUser
-import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.*
 
 /**
@@ -48,7 +47,11 @@ data class RestrictedMemberChatMember(
     @SerialName(canPinMessagesField)
     override val canPinMessages: Boolean = false,
     @SerialName(canManageTopicsField)
-    override val canManageTopics: Boolean = false
+    override val canManageTopics: Boolean = false,
+    @SerialName(canEditTagField)
+    override val canEditTag: Boolean = false,
+    @SerialName(tagField)
+    override val tag: UserTag? = null,
 ) : RestrictedChatMember, SpecialRightsChatMember, MemberChatMember, ChatPermissions {
     @SerialName(statusField)
     @Required